requirements.txt
requests==2.31.0
aiohttp==3.9.1
aiosqlite==0.19.0
python-dotenv==1.0.0
PyMySQL==1.1.0 